+++ Sources/Taps/Tester/DoesNotThrow.swift
@@ -56,7 +56,7 @@ public extension Test {
     line: Int = #line,
     column: Int = #column,
     function: String = #function,
-    test: @autoclosure () throws -> T
+    test: () throws -> T
   ) {
     report.onNext(assertDoesNotThrow(
       message,
@@ -86,7 +86,7 @@ public extension Test {
     line: Int = #line,
     column: Int = #column,
     function: String = #function,
-    test: @autoclosure () throws -> T
+    test: () throws -> T
   ) {
     let location = SourceLocation(file: file, line: line, column: column, function: function)
     report.onNext(assertDoesNotThrow(of: E.self, message, source: location, test: test))